move compound interest type to separate file
